refactor: move registerComponent out of post-process
