+++ packages/@lwc/babel-plugin-component/src/post-process/transform.js
@@ -10,13 +10,10 @@
  * of all transformed decorators.
  */
 
-const { basename, extname } = require('path');
 const moduleImports = require('@babel/helper-module-imports');
+const { REGISTER_DECORATORS_ID } = require('../constants');
 const { isLWCNode } = require('../utils');
 const LWC_POST_PROCCESED = Symbol();
-
-const REGISTER_DECORATORS_ID = 'registerDecorators';
-const REGISTER_COMPONENT_ID = 'registerComponent';
 
 module.exports = function postProcess({ types: t }) {
     function collectDecoratedProperties(body) {
@@ -84,63 +81,7 @@ module.exports = function postProcess({ types: t }) {
         return t.callExpression(id, [klass, t.objectExpression(props)]);
     }
 
-    function getBaseName({ file }) {
-        const classPath = file.opts.filename;
-        const ext = extname(classPath);
-        return basename(classPath, ext);
-    }
-
-    function importDefaultTemplate(path, state) {
-        const componentName = getBaseName(state);
-        return moduleImports.addDefault(path, `./${componentName}.html`, {
-            nameHint: 'tmpl',
-        });
-    }
-
-    function createRegisterComponent(declarationPath, state) {
-        const id = moduleImports.addNamed(declarationPath, REGISTER_COMPONENT_ID, 'lwc');
-        const templateIdentifier = importDefaultTemplate(declarationPath, state);
-        const statementPath = declarationPath.getStatementParent();
-        let node = declarationPath.node;
-
-        if (declarationPath.isClassDeclaration()) {
-            const hasIdentifier = t.isIdentifier(node.id);
-            if (hasIdentifier) {
-                statementPath.insertBefore(node);
-                node = node.id;
-            } else {
-                // if it does not have an id, we can treat it as a ClassExpression
-                node.type = 'ClassExpression';
-            }
-        }
-
-        return t.callExpression(id, [
-            node,
-            t.objectExpression([t.objectProperty(t.identifier('tmpl'), templateIdentifier)]),
-        ]);
-    }
-
-    function needsComponentRegistration(path) {
-        return (
-            (path.isIdentifier() && path.node.name !== 'undefined' && path.node.name !== 'null') ||
-            // path.isMemberExpression() || // this will probably yield more false positives than anything else
-            path.isCallExpression() ||
-            path.isClassDeclaration() ||
-            path.isConditionalExpression()
-        );
-    }
-
     return {
-        // Register component
-        ExportDefaultDeclaration(path, state) {
-            const implicitResolution = !state.opts.isExplicitImport;
-            if (implicitResolution) {
-                const declaration = path.get('declaration');
-                if (needsComponentRegistration(declaration)) {
-                    declaration.replaceWith(createRegisterComponent(declaration, state));
-                }
-            }
-        },
         // Decorator collector for class expressions
         ClassExpression(path) {
             const { node } = path;
